Change the playGame function to work as intended
The function now uses a while loop that checks if any player's score
is less than 5, if it is, it gets a random input from the computer,
evaluates the result and prints in the log.

After any of the players reaching a score of 5, the game stops and
prints the game result with the score of the user.